move the zuora reader only used for test to the test package
